Make router work
