+++ app/ide-desktop/lib/dashboard/src/App.tsx
@@ -38,88 +38,36 @@ import * as React from 'react'
 import * as reactQuery from '@tanstack/react-query'
 import * as router from 'react-router-dom'
 import * as toastify from 'react-toastify'
-
-import * as detect from 'enso-common/src/detect'
+import * as z from 'zod'
 
 import * as appUtils from '#/appUtils'
 
-import * as inputBindingsModule from '#/configurations/inputBindings'
-
-import * as backendHooks from '#/hooks/backendHooks'
-
-import AuthProvider, * as authProvider from '#/providers/AuthProvider'
-import BackendProvider from '#/providers/BackendProvider'
-import * as httpClientProvider from '#/providers/HttpClientProvider'
-import InputBindingsProvider from '#/providers/InputBindingsProvider'
-import LocalStorageProvider, * as localStorageProvider from '#/providers/LocalStorageProvider'
-import LoggerProvider from '#/providers/LoggerProvider'
+import LocalStorageProvider from '#/providers/LocalStorageProvider'
 import type * as loggerProvider from '#/providers/LoggerProvider'
-import ModalProvider, * as modalProvider from '#/providers/ModalProvider'
-import * as navigator2DProvider from '#/providers/Navigator2DProvider'
-import SessionProvider from '#/providers/SessionProvider'
-import * as textProvider from '#/providers/TextProvider'
+import ModalProvider from '#/providers/ModalProvider'
 
-import ConfirmRegistration from '#/pages/authentication/ConfirmRegistration'
-import ForgotPassword from '#/pages/authentication/ForgotPassword'
-import Login from '#/pages/authentication/Login'
-import Registration from '#/pages/authentication/Registration'
-import ResetPassword from '#/pages/authentication/ResetPassword'
-import RestoreAccount from '#/pages/authentication/RestoreAccount'
-import SetUsername from '#/pages/authentication/SetUsername'
-import Dashboard from '#/pages/dashboard/Dashboard'
-import * as subscribe from '#/pages/subscribe/Subscribe'
-import * as subscribeSuccess from '#/pages/subscribe/SubscribeSuccess'
-
-import * as openAppWatcher from '#/layouts/OpenAppWatcher'
-
-import * as errorBoundary from '#/components/ErrorBoundary'
-import * as offlineNotificationManager from '#/components/OfflineNotificationManager'
-import * as paywall from '#/components/Paywall'
-import * as rootComponent from '#/components/Root'
-import * as suspense from '#/components/Suspense'
-
-import AboutModal from '#/modals/AboutModal'
-import * as setOrganizationNameModal from '#/modals/SetOrganizationNameModal'
-import * as termsOfServiceModal from '#/modals/TermsOfServiceModal'
-
-import LocalBackend from '#/services/LocalBackend'
 import * as projectManager from '#/services/ProjectManager'
-import RemoteBackend from '#/services/RemoteBackend'
 
 import * as appBaseUrl from '#/utilities/appBaseUrl'
-import * as eventModule from '#/utilities/event'
 import type HttpClient from '#/utilities/HttpClient'
 import LocalStorage from '#/utilities/LocalStorage'
-import * as object from '#/utilities/object'
-
-import * as authServiceModule from '#/authentication/service'
 
 import type * as types from '../../types/types'
+import * as appRoutes from './Routes'
 
 // ============================
 // === Global configuration ===
 // ============================
+const INPUT_BINDINGS_SCHEMA = z.record(z.array(z.string()).readonly()).or(z.null())
 
 declare module '#/utilities/LocalStorage' {
   /** */
   interface LocalStorageData {
-    readonly inputBindings: Readonly<Record<string, readonly string[]>>
+    readonly inputBindings: z.infer<typeof INPUT_BINDINGS_SCHEMA>
   }
 }
 
-LocalStorage.registerKey('inputBindings', {
-  tryParse: value =>
-    typeof value !== 'object' || value == null
-      ? null
-      : Object.fromEntries(
-          Object.entries<unknown>({ ...value }).flatMap(kv => {
-            const [k, v] = kv
-            return Array.isArray(v) && v.every((item): item is string => typeof item === 'string')
-              ? [[k, v]]
-              : []
-          })
-        ),
-})
+LocalStorage.registerKey('inputBindings', { isUserSpecific: true, schema: INPUT_BINDINGS_SCHEMA })
 
 // ======================
 // === getMainPageUrl ===
@@ -147,8 +95,6 @@ export interface AppProps {
   /** Whether the application supports deep links. This is only true when using
    * the installed app on macOS and Windows. */
   readonly supportsDeepLinks: boolean
-  /** Whether the dashboard should be rendered. */
-  readonly shouldShowDashboard: boolean
   /** The name of the project to open on startup, if any. */
   readonly initialProjectName: string | null
   readonly onAuthenticated: (accessToken: string | null) => void
@@ -182,6 +128,8 @@ export default function App(props: AppProps) {
     },
   })
 
+  const mainPageUrl = getMainPageUrl()
+
   // Both `BackendProvider` and `InputBindingsProvider` depend on `LocalStorageProvider`.
   // Note that the `Router` must be the parent of the `AuthProvider`, because the `AuthProvider`
   // will redirect the user between the login/register pages and the dashboard.
@@ -196,326 +144,16 @@ export default function App(props: AppProps) {
         transition={toastify.Zoom}
         limit={3}
       />
-      <router.BrowserRouter basename={getMainPageUrl().pathname}>
-        <LocalStorageProvider>
-          <ModalProvider>
-            <AppRouter {...props} projectManagerRootDirectory={rootDirectoryPath} />
-          </ModalProvider>
-        </LocalStorageProvider>
-      </router.BrowserRouter>
+      <LocalStorageProvider>
+        <ModalProvider>
+          <appRoutes.Routes
+            {...props}
+            basename={mainPageUrl.pathname}
+            projectManagerRootDirectory={rootDirectoryPath}
+            mainPageUrl={mainPageUrl}
+          />
+        </ModalProvider>
+      </LocalStorageProvider>
     </>
   )
-}
-
-// =================
-// === AppRouter ===
-// =================
-
-/** Props for an {@link AppRouter}. */
-export interface AppRouterProps extends AppProps {
-  readonly projectManagerRootDirectory: projectManager.Path | null
-}
-
-/** Router definition for the app.
- *
- * The only reason the {@link AppRouter} component is separate from the {@link App} component is
- * because the {@link AppRouter} relies on React hooks, which can't be used in the same React
- * component as the component that defines the provider. */
-function AppRouter(props: AppRouterProps) {
-  const { logger, isAuthenticationDisabled, shouldShowDashboard, httpClient } = props
-  const { onAuthenticated, projectManagerUrl, projectManagerRootDirectory } = props
-  const { portalRoot } = props
-  // `navigateHooks.useNavigate` cannot be used here as it relies on `AuthProvider`, which has not
-  // yet been initialized at this point.
-  // eslint-disable-next-line no-restricted-properties
-  const navigate = router.useNavigate()
-  const { getText } = textProvider.useText()
-  const { localStorage } = localStorageProvider.useLocalStorage()
-  const { setModal } = modalProvider.useSetModal()
-  const navigator2D = navigator2DProvider.useNavigator2D()
-
-  const localBackend = React.useMemo(
-    () =>
-      projectManagerUrl != null && projectManagerRootDirectory != null
-        ? new LocalBackend(projectManagerUrl, projectManagerRootDirectory)
-        : null,
-    [projectManagerUrl, projectManagerRootDirectory]
-  )
-
-  const remoteBackend = React.useMemo(
-    () => new RemoteBackend(httpClient, logger, getText),
-    [httpClient, logger, getText]
-  )
-
-  backendHooks.useObserveBackend(remoteBackend)
-  backendHooks.useObserveBackend(localBackend)
-
-  if (detect.IS_DEV_MODE) {
-    // @ts-expect-error This is used exclusively for debugging.
-    window.navigate = navigate
-  }
-
-  const [inputBindingsRaw] = React.useState(() => inputBindingsModule.createBindings())
-
-  React.useEffect(() => {
-    const savedInputBindings = localStorage.get('inputBindings')
-    if (savedInputBindings != null) {
-      const filteredInputBindings = object.mapEntries(
-        inputBindingsRaw.metadata,
-        k => savedInputBindings[k]
-      )
-      for (const [bindingKey, newBindings] of object.unsafeEntries(filteredInputBindings)) {
-        for (const oldBinding of inputBindingsRaw.metadata[bindingKey].bindings) {
-          inputBindingsRaw.delete(bindingKey, oldBinding)
-        }
-        for (const newBinding of newBindings ?? []) {
-          inputBindingsRaw.add(bindingKey, newBinding)
-        }
-      }
-    }
-  }, [localStorage, inputBindingsRaw])
-
-  const inputBindings = React.useMemo(() => {
-    const updateLocalStorage = () => {
-      localStorage.set(
-        'inputBindings',
-        Object.fromEntries(
-          Object.entries(inputBindingsRaw.metadata).map(kv => {
-            const [k, v] = kv
-            return [k, v.bindings]
-          })
-        )
-      )
-    }
-    return {
-      /** Transparently pass through `handler()`. */
-      get handler() {
-        return inputBindingsRaw.handler.bind(inputBindingsRaw)
-      },
-      /** Transparently pass through `attach()`. */
-      get attach() {
-        return inputBindingsRaw.attach.bind(inputBindingsRaw)
-      },
-      reset: (bindingKey: inputBindingsModule.DashboardBindingKey) => {
-        inputBindingsRaw.reset(bindingKey)
-        updateLocalStorage()
-      },
-      add: (bindingKey: inputBindingsModule.DashboardBindingKey, binding: string) => {
-        inputBindingsRaw.add(bindingKey, binding)
-        updateLocalStorage()
-      },
-      delete: (bindingKey: inputBindingsModule.DashboardBindingKey, binding: string) => {
-        inputBindingsRaw.delete(bindingKey, binding)
-        updateLocalStorage()
-      },
-      /** Transparently pass through `metadata`. */
-      get metadata() {
-        return inputBindingsRaw.metadata
-      },
-      /** Transparently pass through `register()`. */
-      get register() {
-        return inputBindingsRaw.unregister.bind(inputBindingsRaw)
-      },
-      /** Transparently pass through `unregister()`. */
-      get unregister() {
-        return inputBindingsRaw.unregister.bind(inputBindingsRaw)
-      },
-    }
-  }, [localStorage, inputBindingsRaw])
-
-  const mainPageUrl = getMainPageUrl()
-
-  const authService = React.useMemo(() => {
-    const authConfig = { navigate, ...props }
-    return authServiceModule.initAuthService(authConfig)
-  }, [props, navigate])
-
-  const userSession = authService?.cognito.userSession.bind(authService.cognito) ?? null
-  const refreshUserSession =
-    authService?.cognito.refreshUserSession.bind(authService.cognito) ?? null
-  const registerAuthEventListener = authService?.registerAuthEventListener ?? null
-
-  React.useEffect(() => {
-    if ('menuApi' in window) {
-      window.menuApi.setShowAboutModalHandler(() => {
-        setModal(<AboutModal />)
-      })
-    }
-  }, [setModal])
-
-  React.useEffect(() => {
-    const onKeyDown = navigator2D.onKeyDown.bind(navigator2D)
-    document.addEventListener('keydown', onKeyDown)
-    return () => {
-      document.removeEventListener('keydown', onKeyDown)
-    }
-  }, [navigator2D])
-
-  React.useEffect(() => {
-    let isClick = false
-    const onMouseDown = () => {
-      isClick = true
-    }
-    const onMouseUp = (event: MouseEvent) => {
-      if (
-        isClick &&
-        !eventModule.isElementTextInput(event.target) &&
-        !eventModule.isElementPartOfMonaco(event.target) &&
-        !eventModule.isElementTextInput(document.activeElement)
-      ) {
-        const selection = document.getSelection()
-        const app = document.getElementById('app')
-        const appContainsSelection =
-          app != null &&
-          selection != null &&
-          selection.anchorNode != null &&
-          app.contains(selection.anchorNode) &&
-          selection.focusNode != null &&
-          app.contains(selection.focusNode)
-        if (!appContainsSelection) {
-          selection?.removeAllRanges()
-        }
-      }
-    }
-    const onSelectStart = () => {
-      isClick = false
-    }
-
-    document.addEventListener('mousedown', onMouseDown)
-    document.addEventListener('mouseup', onMouseUp)
-    document.addEventListener('selectstart', onSelectStart)
-    return () => {
-      document.removeEventListener('mousedown', onMouseDown)
-      document.removeEventListener('mouseup', onMouseUp)
-      document.removeEventListener('selectstart', onSelectStart)
-    }
-  }, [])
-
-  const routes = (
-    <router.Routes>
-      {/* Login & registration pages are visible to unauthenticated users. */}
-      <router.Route element={<authProvider.GuestLayout />}>
-        <router.Route path={appUtils.REGISTRATION_PATH} element={<Registration />} />
-        <router.Route path={appUtils.LOGIN_PATH} element={<Login />} />
-      </router.Route>
-
-      {/* Protected pages are visible to authenticated users. */}
-      <router.Route element={<authProvider.NotDeletedUserLayout />}>
-        <router.Route element={<authProvider.ProtectedLayout />}>
-          <router.Route element={<termsOfServiceModal.TermsOfServiceModal />}>
-            <router.Route element={<setOrganizationNameModal.SetOrganizationNameModal />}>
-              <router.Route element={<openAppWatcher.OpenAppWatcher />}>
-                <router.Route
-                  path={appUtils.DASHBOARD_PATH}
-                  element={shouldShowDashboard && <Dashboard {...props} />}
-                />
-
-                <router.Route
-                  path={appUtils.SUBSCRIBE_PATH}
-                  element={
-                    <errorBoundary.ErrorBoundary>
-                      <suspense.Suspense>
-                        <subscribe.Subscribe />
-                      </suspense.Suspense>
-                    </errorBoundary.ErrorBoundary>
-                  }
-                />
-              </router.Route>
-            </router.Route>
-          </router.Route>
-
-          <router.Route
-            path={appUtils.SUBSCRIBE_SUCCESS_PATH}
-            element={
-              <errorBoundary.ErrorBoundary>
-                <suspense.Suspense>
-                  <subscribeSuccess.SubscribeSuccess />
-                </suspense.Suspense>
-              </errorBoundary.ErrorBoundary>
-            }
-          />
-        </router.Route>
-      </router.Route>
-
-      <router.Route element={<termsOfServiceModal.TermsOfServiceModal />}>
-        {/* Semi-protected pages are visible to users currently registering. */}
-        <router.Route element={<authProvider.NotDeletedUserLayout />}>
-          <router.Route element={<authProvider.SemiProtectedLayout />}>
-            <router.Route path={appUtils.SET_USERNAME_PATH} element={<SetUsername />} />
-          </router.Route>
-        </router.Route>
-      </router.Route>
-
-      {/* Other pages are visible to unauthenticated and authenticated users. */}
-      <router.Route path={appUtils.CONFIRM_REGISTRATION_PATH} element={<ConfirmRegistration />} />
-      <router.Route path={appUtils.FORGOT_PASSWORD_PATH} element={<ForgotPassword />} />
-      <router.Route path={appUtils.RESET_PASSWORD_PATH} element={<ResetPassword />} />
-
-      {/* Soft-deleted user pages are visible to users who have been soft-deleted. */}
-      <router.Route element={<authProvider.ProtectedLayout />}>
-        <router.Route element={<authProvider.SoftDeletedUserLayout />}>
-          <router.Route path={appUtils.RESTORE_USER_PATH} element={<RestoreAccount />} />
-        </router.Route>
-      </router.Route>
-
-      {/* 404 page */}
-      <router.Route path="*" element={<router.Navigate to="/" replace />} />
-    </router.Routes>
-  )
-
-  let result = routes
-
-  if (detect.IS_DEV_MODE) {
-    result = <paywall.PaywallDevtools>{result}</paywall.PaywallDevtools>
-  }
-
-  result = <errorBoundary.ErrorBoundary>{result}</errorBoundary.ErrorBoundary>
-  result = <InputBindingsProvider inputBindings={inputBindings}>{result}</InputBindingsProvider>
-  result = (
-    <AuthProvider
-      shouldStartInOfflineMode={isAuthenticationDisabled}
-      authService={authService}
-      onAuthenticated={onAuthenticated}
-    >
-      {result}
-    </AuthProvider>
-  )
-
-  result = (
-    <BackendProvider remoteBackend={remoteBackend} localBackend={localBackend}>
-      {result}
-    </BackendProvider>
-  )
-
-  result = (
-    <SessionProvider
-      saveAccessToken={authService?.cognito.saveAccessToken.bind(authService.cognito) ?? null}
-      mainPageUrl={mainPageUrl}
-      userSession={userSession}
-      registerAuthEventListener={registerAuthEventListener}
-      refreshUserSession={refreshUserSession}
-    >
-      {result}
-    </SessionProvider>
-  )
-  result = <LoggerProvider logger={logger}>{result}</LoggerProvider>
-  result = (
-    <rootComponent.Root navigate={navigate} portalRoot={portalRoot}>
-      {result}
-    </rootComponent.Root>
-  )
-  result = (
-    <offlineNotificationManager.OfflineNotificationManager>
-      {result}
-    </offlineNotificationManager.OfflineNotificationManager>
-  )
-  result = (
-    <httpClientProvider.HttpClientProvider httpClient={httpClient}>
-      {result}
-    </httpClientProvider.HttpClientProvider>
-  )
-
-  result = <LoggerProvider logger={logger}>{result}</LoggerProvider>
-
-  return result
 }
