+++ app/ide-desktop/lib/dashboard/src/pages/subscribe/Subscribe/index.ts
@@ -3,4 +3,11 @@
  *
  * Barrel file for Subscribe page.
  */
-export * from './Subscribe'
+import defineRoute from '#/utilities/defineRoute'
+
+import * as subscribe from './Subscribe'
+
+export default defineRoute({
+  path: '/subscribe',
+  element: subscribe.Subscribe,
+})
