+++ app/ide-desktop/lib/dashboard/src/pages/subscribe/SubscribeSuccess/SubscribeSuccess.tsx
