+++ app/ide-desktop/lib/dashboard/src/pages/subscribe/SubscribeSuccess/index.ts
@@ -1,0 +1,8 @@
+import defineRoute from '#/utilities/defineRoute'
+
+import * as subscribeSuccess from './SubscribeSuccess'
+
+export default defineRoute({
+  path: '/subscribe/success',
+  element: subscribeSuccess.SubscribeSuccess,
+})
